base distort time

+++ usr/src/minix/servers/pm/mproc.h
@@ -26,7 +26,7 @@ EXTERN struct mproc {
   /* Child user and system times. Accounting done on child exit. */
   clock_t mp_child_utime;	/* cumulative user time of children */
   clock_t mp_child_stime;	/* cumulative sys time of children */
-  uint8_t mp_child_scale;	/* scale result of distort */
+  float mp_child_scale;	/* scale result of distort */
 
   /* Real and effective uids and gids. */
   uid_t mp_realuid;		/* process' real uid */
